Separation of concerns for the posts component, now Profile is only for the header and main layout for other user routes
